refactor(ingest): extract URL-import staging into its own module
The staging cluster behind POST /ingest/upload/url — the joint budget, the
bounded put and its abandonment reaper, the commit seam and the failure
settlement — is 438 lines on one subject, and it is most of what took the
ingest router past 2600 lines. It moves to url_import_staging.py. The route
handler stays, as do _url_import_filename and _raster_stamped_metadata (the
latter is shared with the direct and presigned doors).

Pure move: every moved symbol's token stream, comments and docstrings
included, is identical to the one on main, and the router's only added lines
are the two import statements. _preflight_dns_budget and
_BUDGET_EXHAUSTED_DETAIL travel with the eleven symbols the issue names: the
constant is read by _remaining_fetch_budget, so leaving it behind would mean
importing a private name backwards out of the router, and the issue's own
438-line measurement counts all thirteen.

Patch targets follow each symbol to the module that resolves it, including
the put-site ownership table and the frozen-clock helper, which now pins both
namespaces the flow reads.

Closes #1711
